% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{aggregate_indlevel}
\alias{aggregate_indlevel}
\title{Aggregate individual-level data}
\usage{
aggregate_indlevel(df_data, age_vec = 0:100, t_max = 100)
}
\arguments{
\item{df_data}{dataframe of individual-level data, in the format output by
\code{sim_indlevel()}.}

\item{age_vec}{an integer sequence of ages over which to aggregate.}

\item{t_max}{the maximum time considered when tabulating durations in each
state.}
}
\description{
Given a line list of individual-level data (for example
  generated by \code{sim_indlevel()}), aggregate data by each 1-year age group.
}
\details{
Transmision probabilities can be derived from the number of patients
  going down each pathway (numerator) and the total number of patients
  (denominator) at each stage. These two outputs are given for each
  transition probability. For lengths of stay in each state we require a
  different tabulation in terms of the number of patients that were in a
  given state for x days. This output is given for each state. All outputs
  are given broken down by age.
}
