% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{cubic_spline}
\alias{cubic_spline}
\title{Get cubic spline from node positions}
\usage{
cubic_spline(x, y, x_pred)
}
\arguments{
\item{x, y}{coordinates of spline nodes.}

\item{x_pred}{x-coordinates at which to evaluate and return cubic spline.}
}
\description{
Given a series of x,y node positions, return a cubic spline
  through these points.
}
