% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_data_quantiles_m}
\alias{get_data_quantiles_m}
\title{Get 95\% confidence intervals from data}
\usage{
get_data_quantiles_m(age, max_age, delta)
}
\arguments{
\item{age}{vector of ages for each patient.}

\item{max_age}{maximum age used in tabulation.}

\item{delta}{for each patient, for a given outcome, the number of days
between start and end dates.}
}
\description{
Given line list data, return 95\% confidence intervals of time
  to a given outcome.
}
