% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_data_quantiles_p}
\alias{get_data_quantiles_p}
\title{Get 95\% exact binomial intervals from data}
\usage{
get_data_quantiles_p(age, max_age, outcome1, outcome2)
}
\arguments{
\item{age}{vector of ages for each patient.}

\item{max_age}{maximum age used in tabulation.}

\item{outcome1, outcome2}{for each patient, a logical TRUE/FALSE as to whether
they satisfy outcome1 or outcome2.}
}
\description{
Given line list data, return exact binomial 95\% confidence
  intervals of a given outcome vs the alternative. The numerator in the
  interval calculation is the number of patients outcome1, and the
  denominator is the number of patients in either outcome1 or outcome2.
}
