% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{get_spline_quantiles}
\alias{get_spline_quantiles}
\title{Get 95\% quantiles over posterior splines}
\usage{
get_spline_quantiles(spline_draws)
}
\arguments{
\item{spline_draws}{series of draws of posterior splines, as returned by the
\code{get_spline()} function.}
}
\description{
Given output from \code{get_spline()}, produce 95\% quantiles
  over all draws.
}
