% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_autocorrelation}
\alias{plot_autocorrelation}
\title{Plot autocorrelation}
\usage{
plot_autocorrelation(
  x,
  lag = 20,
  par = NULL,
  chain = 1,
  phase = "sampling",
  rung = 1
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}}

\item{lag}{calculate autocorrelation up to this many lags.}

\item{par}{which parameter to plot.}

\item{chain}{which chain to plot.}

\item{phase}{whether to plot from burnin or sampling phase.}

\item{rung}{which thermodynamic rung to plot.}
}
\description{
Plot loglikelihood 95\% credible intervals.
}
