% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{Plot parameter autocorrelation}
\usage{
plot_cor(
  x,
  parameter1,
  parameter2,
  downsample = TRUE,
  phase = "sampling",
  rung = 1
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}.}

\item{parameter1, parameter2}{which parameters to plot.}

\item{downsample}{whether to downsample iterations.}

\item{phase}{whether to plot from burnin or sampling phase.}

\item{rung}{which thermodynamic rung to plot.}
}
\description{
Plot parameter autocorrelation.
}
