% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_credible}
\alias{plot_credible}
\title{Credible interval plot}
\usage{
plot_credible(x, show = NULL, phase = "sampling", rung = 1, param_names = NULL)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}.}

\item{show}{which parameters to plot.}

\item{phase}{whether to plot from burnin or sampling phase.}

\item{rung}{which thermodynamic rung to plot.}

\item{param_names}{manually define parameter names on plot.}
}
\description{
Produce credible interval plot.
}
