% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_CrI}
\alias{plot_CrI}
\title{Produce credible interval plot from dataframe}
\usage{
plot_CrI(x, show = NULL, param_names = NULL, rotate = FALSE)
}
\arguments{
\item{x}{a dataframe.}

\item{show}{which parameters to plot.}

\item{param_names}{manually define parameter names.}

\item{rotate}{If TRUE, parameter names are rotated.}
}
\description{
Combined plot of posterior summaries of a given parameter or set of parameters.
}
