% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_par}
\alias{plot_par}
\title{Plot parameter summary}
\usage{
plot_par(
  x,
  show = NULL,
  hide = NULL,
  lag = 20,
  downsample = TRUE,
  phase = "sampling",
  rung = 1,
  display = TRUE
)
}
\arguments{
\item{x}{an object of class \code{drjacoby_output}.}

\item{show}{which parameters to plot.}

\item{hide}{which parameters to omit from plot.}

\item{lag}{calculate autocorrelation up to this many lags.}

\item{downsample}{whether to downsample iterations.}

\item{phase}{whether to plot from burnin or sampling phase.}

\item{rung}{which thermodynamic rung to plot.}

\item{display}{if FALSE the plot is produced but not saved.}
}
\description{
Combined plot of posterior summaries of a given parameter or set of parameters.
}
