% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spline_quantiles}
\alias{plot_spline_quantiles}
\title{Plot spline quantiles}
\usage{
plot_spline_quantiles(
  df_spline_quantile,
  df_data_quantile = NULL,
  title = "",
  ylim = c(0, 1)
)
}
\arguments{
\item{df_spline_quantile}{dataframe of quantiles broken down by age, as
produced by \code{get_spline_quantiles()}.}

\item{df_data_quantile}{dataframe of exact data quantiles, as produced by
\code{get_data_quantiles_p()} or \code{get_data_quantiles_m()}. Ignored if
\code{NULL}.}

\item{title}{plot title.}

\item{ylim}{y-limits of the plot.}
}
\description{
Given age-dependent quantile information, as produced by
  \code{get_spline_quantiles()}, produce a ribbon plot over ages. Optionally
  overlay
}
