% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run main MCMC}
\usage{
run_mcmc(
  data_list,
  df_params,
  burnin = 1000,
  samples = 10000,
  beta_vec = 1,
  chains = 1,
  n_threads = 1,
  pb_markdown = FALSE,
  silent = FALSE,
  return_fit = FALSE
)
}
\arguments{
\item{data_list}{List of data in defined format (see implementation scripts).}

\item{df_params}{Dataframe of parameters in same format as drjacoby package.}

\item{burnin}{Burn-in iterations.}

\item{samples}{Sampling iterations.}

\item{beta_vec}{A vector of powers that allow for thermodynamic MCMC. If set
at 1 then thermodynamic MCMC is effectively turned off and this simplifies
to ordinary MCMC.}

\item{chains}{Independent MCMC chains.}

\item{n_threads}{Number of threads to use if running in
parallel. Do not exceed the number of cores available on the
machine. Has no effect if markovid was not compiled with openmp
support.}

\item{pb_markdown}{If TRUE then run in markdown safe mode.}

\item{silent}{If TRUE then console output is suppressed.}

\item{return_fit}{If TRUE then MCMC is not run, instead the model fit is
returned under the initial parameter values specified in the parameters
dataframe.}
}
\description{
Take in data, model parameters, and MCMC parameters. Run main
  MCMC inferring model parameters using the likelihood within this package.
}
