% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{sim_indlevel}
\alias{sim_indlevel}
\title{Simulate individual-level data}
\usage{
sim_indlevel(params_scalar, params_age, age_vec, df_sim)
}
\arguments{
\item{params_scalar}{vector of scalar parameters.}

\item{params_age}{list of age-varying parameters, each of which is a vector
over integer ages.}

\item{age_vec}{the vector of ages corresponding to the age-varying
parameters.}

\item{df_sim}{dataframe of basic data properties. Must include the columns
"date_admission", "date_censor" and "age".}
}
\description{
Draw individual-level data at random from the basic hospital
  progression model.
}
